A first rewrite of the script.
Remove some hardcoded username in a path. The script still expects the
existence of some specific folders, which should be changed, i.e. they
should be set via a config file or during a configuration stage.

Remove some code doubling. The if clause now only sets runcommands and
prints a message what is/will be done. Note that this means that for the
parallelized case of switch_reconstruction.sh 1 now the code is run in
parallel. So far the code was run with only one processor, i am not sure
if this is a copy and paste error or if this was intended/required.
There is still some code doubling that could be avoided, as the
different runs differ just by a number.

Use variables for the commandline arguments. The commandline arguments
are written to variables, and these are used throughout the script.