Warning if #procs != #spec for multispecies run.
Issue a warning if the number of processors is not the same as the
number of species for a multispecies run.
This may lead to problems, as at some points in the code this equality
is assumed. One known issue is that the collision matrix is not fully
computed/initialized in this case.