Modifications in neo_sub so both files are equal.
Only two points where necessary.
First in neo_init the calculation of rt0 and bmref had to be unified.
For this purpose the recently introduced module global_paramters is used
to make the initialization depending on which code is calling the
subroutine.

Secondly, in neo_read a case had to be ported (copied) to the parallel
code.