Add comment as reminder to check something.
The array anumm_inf was reported by valgrind for uninitialized use. Bug
report was about writing the array to a hdf5 file, but i am not sure if
not only the first use is reported, i.e. if the array might not be used
uninitialized elsewhere.
As at the momement i have not much knowledge of the code, just a todo is
added.