Add script to read/plot in efit files.
Add a python3 script.
It defines a class, whose init method takes an filename for an efit(?)
file to read in and store the information in members.

Note that so far not all the information are processed, as for some
variables/values it is not known, what they are.

If the file is executed as a script, a default file (located in the
shared folder) is used and some of the information from the file are
plotted. The plots are saved to files.