Move mass constants to math_constants.
Collop_compute defined some variables (used as parameter, without the
keyword) for masses of electrons, protons, etc. which are moved to
math_constants. An additional constant ev is renamed to ev_to_cgs, as it
actually converts a value in eV to the energy in cgs units.
Improve the documentation of the constants, including some that
where already in the file.

Also: remove *_D constants, these are not in use, and the normal ones
are of the same type and thus should be good enough.