Add two scripts now required by efit.py.
Booth are originally by Christopher Albert but were modified.
In both cases a if __name__ statement is been added.
In the case of getHeaderDataVMEC.py this retains the original
functionality, but the main part is now in the function
getHeaderDataVmecNc, which returns the data that is printed in the case
of main.
The if __name__ statement in boozer.py mimics the behaviour of a script
that is used to call the original version, which only writes one block
of s-data to a file. The script is responsible for combining the blocks.
Here four functions have been defined. One calculates a block and
appends it to a file, while the other three write certain parts of the
boozer file, the header, header for a block and the block itself.

The latter three functions are also used by the new functions in
efit.py.