Define default values as 'parameter'.
Four default values in mag_interface are now declared with the
'parameter' keyword.
Also the default name is no longer defined with a specified length, so
the length of the constant should be taken. This was done to simplify
the code and avoid warnings because of size missmatch, not so much to
save memory.