Small fixes to avoid floating point exceptions.
Make changes to avoid floating point exceptions during integration.
The first two are regarded to the functions G and d_G, which will now
return for x=0, the limit, instead of trying to compute it (which would
result in a 0/0).
Secondly, the integration of I_phi_1_param  inside of the kernel of
compute_I2_mmp_s (called I_phi), makes trouble for x = 0, thus in this
case 0 is used.
Third also the integration of I_phi_2_param makes problems. One was at
x=0, which resulted in 0/0. This has been fixed in the same manner as
for I_phi_1_param. Just setting the integral to zero is justified by two
points. One if param is zero then for all values of x!=0 the kernel is
zero and the single point should not change the integral (mathematically
not okay, as the value at this point is not determined but as limiting
value it should also be okay mathematically). Second, for the not
stabilized case, the value of the integral is multiplied with x**l, so
the value used further should again be zero for x=0.