Add comparison of hdf5 files.
The test-script now also has and uses a function for comparing the hdf5
files.
In contrast to the *.dat files, only the final hdf5 files are compared.
This is because of difficulties to retain the intermediate hdf5 files.
It is not expected, that this is a serious limitation, as the final
stage depends on all the other stages and most results written to a hdf5
file should be in the final files.