Fix hdf5-check. Ignore objects expected to differ.
Fix of the check of the hdf5-files.
The check will now ignore objects that are expected to differ.
At the moment this is the metadata object of neo2_config and final.
Unfortunately it does not seem to be possible to exclude the metadata
objects in the files with a single exclude-path option, as the specific
path has to be given, which differs in the files (final.h5 gets all the
content from neo2_config?).