Special treatment for parallel case.
The parallel version of the test gets now a special treatment.
For this purpose the additional variable 'accuracy' is introduced. It is
empty for the single processor case. In the case of a parallel run, the
value is set to '--relative=1.0e-13', i.e. only if the relative
difference of two values is bigger, they will be considered as different
by h5diff. Additionaly there is another exclude path, for a field that
is set to 1 for parallel runs.

With this, the hdf5 checks should pass, with one and with more
processors. But for the later case, there are still differences in the
dat files.

With this commit also the output in the case of differences has been
improved. If a check for a file fails, it is now reported which files
where compared, and what where the differences, as h5diff outputs
normally. (You may need to use cmake --verbose to see this output.)