The previous version did not work as desired.
The previous version worked for checking agains the same codeversion,
but when changing the reference to the output of a previous version
(4249957b68), it was not good enough.

Thus some changes.
Accuracy is now used also for serial runs, the variable does not include
the option anymore, and its value has been increased to 1.0e-12.

If a new variable is set to yes (so far it is not changed in the code),
then if the relative test fails, additionaly a test with absolute errors
is made with the same accuracy. Only if this fails too, the differences
are counted as 'real' differences.
The differences for the relative error are nevertheless reported and
there is a message if the second test passes (not if it fails so far,
but first message should be enough).