Minor change to simplify keeping tests up to date.
When copying the template input files, follow links.
Use case is thought to be the following: setting up a reference, then
creating the template directory and linking to the needed files of the
reference folder. If the files need to be updated (should not be that
often), then the template directory is automatically up to date.
Another, maybe even more important point, is to prevent the reference
from getting changed, when using links.