Add copy of external UseLATEX.cmake to ease acess.
Add a copy of the cmake script UseLATEX (I would add here a refence to
where I found it, but I forgot). While there is no intention to develop
this within the group, it is intended to be used for making the
documentation, and as it is not usally available, for easy use, also
distribute this with neo2.