Add document describing the intended coding style.
Add LaTeX document, which describes the coding style that should be used
for new code or when reworking old code.

If someone disagrees with these guidlines, they are open for discussion.

Also note, that these describe the desired state of the code, not the
is-state. That certain parts do not fullfill the guidelines should be no
excuse to ignore them for new code.

External documents (e.g. UseLATEX.cmake), might be exempted from the
guidelines, but ideally they would also follow them.