Add a cmake configuration file for DOC folder.
Add a cmake configuration file for the DOC folder, which so far can make
the doxygen documentation and the coding style document.
As for the PAR and QL variant the makefile is intended to be used with a
subfolder (called, e.g. Build).
The html folder and the coding style pdf will remain in the subfolder,
while the doxygen pdf will be copied into the DOC folder. (This
behaviour should be changed, but for now it is good enough.)

For the creation of the coding style document the UseLATEX.cmake script
is used. More documents might be added easily in a similar way.