Add example script to read data from neo-2 scan.
Add example script to show how one can read the data from a scan neo-2,
using octave (based on a matlab version).
Defines some functions, that might be of more general interest.

Note that the h5read function should not be used with large hdf5
files, as it will read in the whole file, and then select the
requested part, i.e. for large files much memory is needed. Not sure
how the size scales with representation in memory, but I would guess
at least one to one, with tendency to larger memory representations.

Usage of this script requires to set up the scan folders with specific
name structure.