Add basic structure for user manual.
It is intended as follows: the main file includes the files for the
different chapters (so far Introduction and Output).
The file Commands.sty contains the definition of commands (so far only
two commands for first and second partial derivative). This is done so
other documents (e.g. from ExtraDocumentsTex) can use this file (via
\usepackage{})
In CMakeLists.txt is already a rule for making the documentation, but is
has to be adapted, if additional chapters are added.

What is so far not done, but should be done, is to copy the output file
(i.e. the pdf) from the build directory to the DOC directory.