Bugfix: need also to write namelist.
This relates to the introduction of neoin as a namelist. When making
this change I forgot to take into account, that the namelist may also
need to be written.
This is now fixed. The subroutine got an optional switch for this, which
if true, will make the subroutine to write the namelist to the file
'neo2.in' instead of reading it.
It is made sure, that the file is only opened if it is not already
attached to a unit (which would be a bug).

Attention: the name of the input file is still hardcoded, this should be
changed. Best solution would probably be passing it as an argument to
the subroutine.