Add shell script for collection hdf5 data.
This is more or less a reimplementation of h5merge_multispec.x from the
MULTI-SPEC-TOOLS subdirectory, as a shell script.
Major differences, the shell script requires two input arguments, it
does not use find for determining the list of files, and it uses a
program (h5copy) that comes with the hdf5 library to do the job.