Change in mode for testing hdf5 files.
Before this commit, if a relative comparison failed, a test for absolute
differences was done. This second test is now switched off and instead
there will be determined and printed how big the differences are. Test
is done with integer exponents from -12 up to 1.

Testcases now fail with differences up to order of 1 (not exponent), in
/ailmm and /ailmm_aa of collop.h5.