Keep attributes when collecting from subfolders.
Attributs are now also copied from the file in the first subfolder,
assuming that these are the same for all files.
This is done, to keep espacially the comments and units.
Some fields have attributes 'lbound' and 'ubound' which should probably
not be copied, but for simplicity this is ignored for now.