Add function to resize the arrays of a hdf5 file.
Added a function to hdf5tools, that allows to resize (some) arrays in a
hdf5 file.
Main application in mind was resizing a radial hdf5 input file for a
scan, e.g. to exclude the axis from a radial scan, outer points where
the temperature get negative, or points where it is known, that there is
no drive by a mode.