Bugfix: pass self to member and start at zero.
Pass self as argument to member functions. (Did not test if this is
really a bug, but I decided to treat it as such.)
Start loop at zero, and do not ignore the header, as at least one line
of the header contains information we need (regarding the flux surface,
e.g. boozer s).