Add a first version of a class for condor files.
Add a first version of a class for a condor submit file.
Files can already be read and writen, some major values can be set.
Main todo that remains at the moment is the implementation of the queue
command. A version for submitting a list of files is required, as well
as the means to change between the variants.