Rename function and add another.
Rename function to 'remove_ends_from_hdf5_file'. The old name is now
used for a new function, which takes a list of the elements to keep,
instead of a 'range' (first and last element).
The new function is now called by the old function.
The changes required changing the case-functions as well.