Add write method to efit_type class.
The efit_type class has now a write routine, which will write the data
of the class to a file with given name.
Due to poor understanding of the format, the write routine might not
work in all cases, as it it now.
Also the method does so far only write those data that is also read,
i.e. there is some sizable part at the end that is not written.