Bugfix and adaption for reading in efit data.
Bugfix: the two datafields at the end of the efit file would be read
into the same array (copy-paste error?), which might have caused the
problems when trying to read in the second array.

Adaptions:
- the last(?) datafield contains ten more elements (with two
values each, i.e. an additional 20 numbers). The code has been adapted
to this.
Note that it is not known, if this is correct or if the files seen so
far are ill formated.
- write the field also to file.