Move code to subroutine, to improve structure.
Move the code for creating the folders and input files for a
multi-species scan to an own subroutine, to improve the structure and
readability of the code - now it should be much easier to see that the
code stops after this task is done.

A disadvantage may be, that the subroutine makes heavy use of variables
defined in the main program.