Add function to save poloidal cut of flux surfaces
Add a function to BoozerFile class, which will store data for flux
surfaces of a poloidal cut to file, so the data can be plotted, e.g.
with gnuplot.
Intended for checking island width, and in this context as addition or
replacement for surfaceplotter.x (fortran).
Seems to be slower (not measured), and the innermost flux surface has a
radius about 50% larger, for so far not known reasons.