Add script with non-hdf5 functions/tools.
Add a python script with functions/tools not related to hdf5.
So far there is a function for getting the memory consumption from a
series of runs, all of which share an id (a 'batch'?).

It might turn out to be to crude to be of much use and thus requires
rewritting/extension.