Shell/matlab scripts for postprocessing by Matyas.
Add the shell/matlab scripts for postprocessing from Matyas.
On these the recently added python variants are based.

Note that the script 'matyas_postproc.sh' requires the executable
'h5merge_multispec.x' which is not part of this repository (not sure
where this comes from).

Hope I did not miss a necessary file.