Do not create additional folder in testfolder.
So far the reference folder was copied into the testfolder, i.e. inside
the testfolder was just a single folder, which then contained the data.
This additional level has been removed, the data is now directly in the
testfolder. The recent change in the name makes it possible to identify
the testcase.