Fix for ordering of eigenvalues on behalf Sergei.
Fix comes actually from Sergei, but he does not use git.

Ordering of the eigenvalues was only based on magnitude. This caused
problems with checking for convergence, when complex conjugate pairs
were among the eigenvalues, as they could appear in random order in
subsequent iterations.
Now complex pairs are ordered such, that the one with imaginary part
larger than zero will be first in the list.

Attention: this fixes only ArnoldiOrder2, the other two ripple solver of
the ql version are not fixed.