Move three subroutines to module flint_mod.
They were the same in both versions, except for the use of '=' instead
of '=>' for a pointer initialization in the PAR version of sort_theta.
This is assumed to the wrong, i.e. the version with '=>' is now used.
Note that these functions are not called at the moment. There call is
commented in neo2.f90.