Move subroutines to module flint_mod.
They were the same in par and ql version, so this is to avoid code
doubling.
Note that the subroutine lagrange_coefs5 is not used at the moment, as
the call statement is commented. This means there is no use statement
for this subroutine, it needs to be added when using it (e.g. by
uncommenting).