Change regularization of matrix, change at end.
On behalf of Sergei: change the way the regularization of the matrix is
done. Now the compuation of the solution is done, and then the solution
is changed such, that there is no density produced, i.e. this means the
offset created by the algorithm, because the matrix is singular, is
removed.