Add octave scripts for figures for documentation.
Add an octave function and to octave scripts, meant for creating figures
for the documentation (there are not used there at the moment).
One script is for the documentation of the subroutine
'customize_eta_levels', the other for the input parameter
'collop_bspline_dist'. The latter script makes use of the function
'compute_knots'.