Add part for input parameter collop_bspline_dist.
Add paragraph to show the impact of the parameter on the distribution of
points in velocity space.
The image is not yet included, first i have to figure out how to create
it via the make/cmake. It can already be created manually, as the
required octave script is already included.