Add sanity checks for transport coefficients.
Add a first version of a sanity for the computed transport coefficients
related to the particle flux.
They are four derived from the condition of the particle flux beeing
ambipolar and one from the condition that the fflux is independent of
the radial electric field. The latter holds only if centrifugal
effects can be neglected.