Turn into function with old behaviour as default.
Turn the script generate_neo2_profile into a function (of the same
name).
The strings that are function parameters (output filename, path to the
shoot, and shoot designation) will be set to the values set in the
script, i.e. when called without arguments, the function will behave the
same as the script.