Rewrite of matlabplot_allm.
On behalf of Sergei.

Now stores the data for all species in subfolders with species number,
if they exist. If a particular folder does not exist then no output for
that species will be produced.
Output of the grid is now controlled via parameter.
A prefix (2 characters) needs to be given.

Adapt also the call in test_conservation.