Abort if processors unequal species.
Upgrade the check, if the number of processors is equal to the number of
species.
The warning is kept, but additionally, the code will stop if the number
of processors is larger than one.
The case with more than one species but only one processor is kept for
testing purposes.