Bugfix in script.
Used the parametername as filename instead of using the value of the
paramter.

Note: the script is intended for octave and will not work with matlab.
Matlab requires use of h5read, which additionally requires the name of
an entry to load as parameter. The structure must be build on its own.