Scale number subintervalls for integration a to b.
Scale the number of subintervalls for the integration from a to b with
the mass ratio of the species (m_b/m_a, but at least one), in an effort
to improve the result of the integration for large mass ratios (e.g.
electrons-tungsten).
Does improve the result but there are still some suspicious things, e.g.
jumps in K and in subintegrand I1 of part I2.