Table for inner integration.
Via hardcoded switches, the inner integration for matrix elements
ailmm_aa is now computed once and stored in an array. The values for the
required points are calculated from these via interpolation, here the
gamma factor is taken into account.
Some elements required special treatment to avoid Infinity as a result
of the integration.
Should be faster, as the inner integrals are now only computed once.