Bugfix: add missing factor n_spec.
The ambipolarity conditions for the matrix elements where missing a
factor n_spec(k-1), which was somehow missed in the derivation of the
equations.
Note that this was not the case for the condition of the particle flux
for each species to be independent of the radial electric field. For
this it was not forgotten, but here it also would make no difference, as
it would just appear as a scaling factor.

This bug is also fixed in the documentation.