Change namelist read-check functions.
Change the functions for checking/writing the status of reading a
namelist, to allow for optional namelists.
This is done via optional parameter 'required_' which is assumed to be
true if not given.
If a namelist is optional an error during read only produces a warning
and no abort of the program.