Add new function for creating perturbation file.
On behalf of Sergei.
Add new function for creating a boozer file with perturbations of the
magnetic field.
The new function is based on the former one, but changes not only radial
position, also theta position is changed. Change in theta is calculated
from change in radial direction by condition that the change is
divergence free (?).

Note: found some variables which are now basically unused and removed them.
There might be more that I did not notice.