Bugfix: creation of output file did not work.
Bugfix: the creation of the output file did not work as intended (was in
the wrong folder and not only created if necessary).
This should now be fixed, i.e. now the 'postproc_torque' function will
create 'final_neo2_multispecies_out.h5' for each subfolder where it not
already exists.