Change behaviour of hdf5 comparison.
If the number of keys in a group differs, this is no longer considered
to mean the files differ (and thus just worth a warning), as it could be
that new fields have been added which are not present in the reference.
The other way round, that the reference has fields not in the test is
still considered to mean the files differ.

The message now states 'Warning' and has been made more clear what
differs, and where. The latter by adding the name of the parent group to
the output.