Bugfix: Error in determining if keys agree.
Fix a bug in the check if the keys in two hdf5 files agree.

For files with nested groups false positives could appear, because
later checks overwrote the return value of earlier checks, i.e. if the
last check was succesful then the keys are assumed to agree.
This is now fixed, by using  and-operator with the current value.