Add file from Gernot(?) for conversion of profile.
Add matlab file from Gernot(?) for conversion of a specific profile file
format to one suitable for use with neo-2 codes.
(I think this is not used directly by neo-2, just by scripts/codes that
create subfolders for simulations.)

Note that this was originally not a function, and that according to a
statement in Gernots 2018 Documentation, it is only suited for a
particular profile file format.