Add file from Gernot(?) to convert file format.
Add an octave file from Gernot(?) to convert a profile file to TRAVIS
format.

Note that this was originally not a function.

Note that this seems very similar to convert_prf.m, only the conversions
done seem to differ. Thus it might be a good idea to either introduce
some more functions to do the common stuff, or to use one functions for
both and e.g. use a switch to decide between the possibilities.