Remove questions from script.
Question for relativistic part seems unneccessary, as it will do nothing
if the sequence to replace is not present. If it is present in a
non-relativistic run, the parameter should be unused in the code, but if
it is not a number it should lead to an error when trying to read the
namelist.

Question wether to create surfaces was deemed unneccessary, as this is
the whole purpose of the script.