Add parameter to function, change parametername.
Add a parameter to function for copying everything or just a group with
name matching subfolder.
Name of other parameter has been changed, as the old name seemed
confusing (could not figure out what was it intended for, but have only
seen the documentation of the parameter itself, not the explanation in
the text).

The documentation has been adapted accordingly.