Default values for Neo2File class.
The Neo2File class should now have all parameters, including somehow
usable values, although not necessarily the default values used by
neo-2.

To achieve this, a new member variable was added: _variant, and loading
was changed.

The new parameter is passed to the constructor, and should correspond
to one of the keys of the new dictionary example_files. It determines
which of the complete example files (in the DOC folder) should be used.

Loading is now a two step process. In the first process the neo2.in file
to load is put into a f90nml namelist object. In the second step the
example file indicated by _variant is loaded with the namelist object
as patch, i.e. the values will replace the 'defaults' from the example
file.

Also: Documentation of compare function.