Add python function like prop_reconstruct_3.
Add a python function that reimplements subroutine prop_reconstruct_3 of
the PAR version of neo-2.
So far there is one addition, one drawback and one difference.
It has added capability of selecting output file name.
First drawback is reseting of tag_first and tag_last is not implemented.
The difference is that the 'lsw_save_' switches are ignored. The
function checks if the files are present, and if so adds the data and
deletes the files.