Add second version prop_reconstruct_3.
This version has just one dataset for each field, instead of one per
propagator.

To implement this a bunch of helper functions have been written, which
are also added.

Hope was that this might produce smaller final.h5 files than the first
version, but this is not the case. As the size of the dataset for each
propagator differ, and variable length arrays have not been used. (Just
found out about those.)