Add tool create_surfaces.
Add the two tool from /proj/plasma/Neo2/Interface/Create_Surfaces/ to
neo-2 repository, as it was not under version control so far.

The main files of the tools are create_surfaces.f90 and nfp.f90,
respectively.

For now, a pass was done to move each module to an own file, and
subsequently to remove unused files, but this was not exhaustive.

For some, or even most of the files the neo-2/COMMON or libneo
version should be used.

Coding style was not yet modified, i.e. the code does not comply to
neo-2 standards.