Avoid explicit references to /proj/plasma/.
Most of them (Travis section still refers to it explicitly), have been
hidden behind NEO2DATAPATH, with note at the beginning, that this refers
to /proj/plasma/Neo2/ at the ITCP of TU Graz.

One reference to a subfolder of Gernot in /proj/plasma/ was removed as
the scripts there should now mostly have a copy under version control in
neo-2.