Switch behaviour depending on verbostiy.
Up to now, function used short-circuited evaluation every time (which
means here, that subgroups are only compared until a difference is
found).
Now, this is no longer the case if verbose is set to true. Reasoning is
that this is done to find out what differences there are (opposed to if
there are differences), and the short-circuited operation is contrary to
this.