Fix three bugs that prevented compilation.
Introduced during cleanup of module use (1f6fdfb3a1).

In one subroutine of binsplit, 2003 version of hdf5_tools module needed
to be used, as subroutine makes use of one of the subroutines, that are
not in the 'normal' hdf5 module.

In propagator a use statement for hdf5_tools was missing.

In ripple_solver_ArnoldiOrder2 a variable could no longer be used
through a different module.