Special treatment of case when removing species.
Special treatment for the case of two remaining species in the function
for removing species from an neo-2 input file.
To have charge neutrality density, and gradient, of electron and ion
temperature should be the same.

This can be seen only as a temporary solution. It should be changed to
rescale the density of all remaining species.