Add function to rescale quantities in hdf5 file.
Add a function to rescale multiple quantities in a hdf5 files at once
(with the same scaling factor).

The function to rescale a single quantity now makes use of the new
function. (Not the other way round, because that would mean the file
would have to be rewritten for each quantity that is changed.)

With this also the function to rescale density is obsolete, while easier
to use, as one does not need to remember the exact name of the fields.