Add CMake file for COMMON folder.
Add a CMake file for subfolder COMMON.
The files has been succesfully tested with CMakeLists.txt from
NEO-2-PAR, but the correponding code to use it from there is not yet
added.
Also the file neo_sub.f90 and the folder unit_tests are so far missing
from COMMON/.
Some things need to be clarified before they can be added.