Neo-2-par version now makes use of common library.
The neo-2-par version will now use the cmake file in the COMMON folder
and the library created.

To be able to do this neo_sub.f90 has been moved to the COMMON folder.

The common folder itself is added as a link to the NEO-2-PAR/ folder.

Done to (1) have a more modular structure of the build process, and (2)
to be able to create unit-test, which now can use the library.

Note: no changes so far for neo-2-ql, as there are still differences in
the neo_sub.f90 version.