Move ProjectConfig from NEO-2-QL to COMMON folder.
This is an intermediate step, so the code will not compile.

Move files ProjecConfig.cmake.in(.intel) from the NEO-2-QL folder to
the COMMON folder.
Leaves project in non-compilable state, which is done to move the files
and change it in the same commit (reasoning that this would make diffs
harder to read).