Add one input parameter and one output parameter.
Add an input parameter for controlling if large values (>1.0e+10) should
be ignored. Default value used is false, which corresponds to previous
behaviour.
Add logical matrix used for determining region over which to integrate
to list of output variables, as it now might be required to determine
the grid for additional quantities.