Replace local spline_cof.f90 with COMMON variant.
Replace variant of spline_cof.f90 for tool create_surfaces with the
variant from COMMON.
This required some more changes in the CMakeLists file, as some
additional libraries are required.

Note that this caused some differences in a test. The differences are of
relative order 1.0e-14. Also, the differences are not in the flux
surface label itself. This can be explained due to the change in solver
for the spline coefficients.