Remove empty comment lines and commented code.
Remove empty comment lines (not those part of a comment block for
formatting) and commented code. The latter means for example
write/print statements for debugging(?) that was not removed before the
commit.
Some cases where it was not clear if this counts as comment or as code
are left. One example is in magnetics.f90, with comments asking
'how does delete of file now work in modern fortran'.
Another example is counting of matrix elements in ripple solver. The
comment terms are not removed, as they might serve as indication what is
counted. (This should be removed/avoided at some point, as this is a
case of code doubling.)